refactor: extract Axes and InsetAxes into axes/ subpackage
Moves Axes and InsetAxes out of figure_plots.py into anyplotlib/axes/_axes.py
and anyplotlib/axes/_inset_axes.py respectively. Also relocates _plot_kind()
to _inset_axes.py where it is used. Backward-compat shims in figure_plots.py
preserve all existing import paths.
