+++ anyplotlib/axes/__init__.py
@@ -1,0 +1,6 @@
+"""anyplotlib.axes — grid-cell and inset axes containers."""
+
+from anyplotlib.axes._axes import Axes
+from anyplotlib.axes._inset_axes import InsetAxes
+
+__all__ = ["Axes", "InsetAxes"]
